Api/requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astor==0.8.1
botocore==1.20.46
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
Django==3.2
django-storages==1.11.1
djangorestframework==3.12.4
gast==0.2.2
google-pasta==0.2.0
grpcio==1.37.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.10.0
jmespath==0.10.0
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
networkx==2.5.1
numpy==1.16.4
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.15.7
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.6
scikit-image==0.17.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
sqlparse==0.4.1
tensorboard==1.15.0
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==1.1.0
tifffile==2021.3.31
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
channels==3.0.4
channels-redis==3.3.0
daphne==3.0.2
